refactor : UserPageController -> PostPageContorller
