+++ src/main/java/com/woowacourse/zzinbros/post/web/PostPageController.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzinbros.user.web.controller;
+package com.woowacourse.zzinbros.post.web;
 
 import com.woowacourse.zzinbros.post.domain.Post;
 import com.woowacourse.zzinbros.post.service.PostService;
@@ -8,27 +8,27 @@ import com.woowacourse.zzinbros.user.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
 @Controller
-@RequestMapping("/user-page")
-public class UserPageController {
-    private UserService userService;
-    private PostService postService;
+public class PostPageController {
 
-    public UserPageController(UserService userService, PostService postService) {
-        this.userService = userService;
+    private final PostService postService;
+    private final UserService userService;
+
+    public PostPageController(PostService postService, UserService userService) {
         this.postService = postService;
+        this.userService = userService;
     }
 
-    @GetMapping("/{id}")
-    public String showPage(@PathVariable("id") final Long id, Model model) {
+    @GetMapping(value = "posts")
+    public String showPage(@RequestParam("author") final Long id, Model model) {
         try {
             User author = userService.findUserById(id);
             List<Post> posts = postService.readAllByUser(author);
+            model.addAttribute("author", author);
             model.addAttribute("posts", posts);
             return "/user-page";
         } catch (UserNotFoundException e) {
