+++ src/main/resources/templates/element/right-sidebar-detail.html
@@ -1,4 +1,4 @@
-<div th:if="${session.loggedInUser}" class="card">
+<div th:if="${session.loggedInUser}" class="card" xmlns:th="http://www.w3.org/1999/xhtml">
     <h5 class="card-title mrg-top-15 mrg-btm-0 mrg-horizon-15">스토리</h5>
     <div class="card-body">
         <ul class="list-unstyled list-info">
