+++ src/main/resources/templates/layout/center-container.html
@@ -1,4 +1,4 @@
 <div class="col-lg-7 col-md-7 col-sm-7 col-xs-7" xmlns:th="http://www.thymeleaf.org">
     <div th:replace="layout/social-feed"></div>
-    <div th:replace="layout/cards"></div>
+    <div th:replace="element/cards"></div>
 </div>
