+++ src/main/resources/templates/layout/right-sidebar.html
@@ -1,2 +1,3 @@
 <div class="col-lg-3 col-md-3 col-sm-3 col-xs-3 pdd-right-0" xmlns:th="http://www.w3.org/1999/xhtml">
+    <div th:replace="element/right-sidebar-detail"></div>
 </div>
