+++ src/main/resources/templates/user-page.html
@@ -13,6 +13,7 @@
         </div>
     </div>
 </div>
+</div>
 <th:block th:include="common/cdns"></th:block>
 <script src="/js/fetch-api.js"></script>
 <script src="/js/user-page-add.js"></script>
